+++ az/outputs.tf
@@ -9,6 +9,16 @@ output "lan_id" {
   value = "${join(",",aws_subnet.lan.*.id)}"
 }
 
-output "nat_eip_id" {
-  value = "${join(",",aws_eip.nat_eip.*.id)}"
+## Returns information about the NATs
+output "eip_nat_id" {
+  value = "${join(",",aws_eip.eip_nat.*.id)}"
+}
+
+output "nat_id" {
+  value = "${join(",",aws_nat_gateway.nat.*.id)}"
+}
+
+## Returns the routing table ID
+output "rt_lan_id" {
+  value = "${join(",",aws_route_table.rt_lan.*.id)}"
 }