+++ base/main.tf
@@ -3,8 +3,10 @@
 ## Provisions Virtual Private Cloud (VPC)
 resource "aws_vpc" "vpc" {
   cidr_block = "${var.vpc_cidr}"
+  instance_tenancy = "${var.instance_tenancy}"
   enable_dns_support = "${var.enable_dns}"
   enable_dns_hostnames = "${var.enable_hostnames}"
+  enable_classiclink = "${var.enable_classiclink}"
 
   tags {
     Name = "${var.stack_item_label}-vpc"
@@ -28,11 +30,6 @@ resource "aws_internet_gateway" "igw" {
 resource "aws_route_table" "rt_dmz" {
   vpc_id = "${aws_vpc.vpc.id}"
 
-  route {
-    cidr_block = "0.0.0.0/0"
-    gateway_id = "${aws_internet_gateway.igw.id}"
-  }
-
   tags {
     Name = "${var.stack_item_label}-dmz"
     application = "${var.stack_item_fullname}"
@@ -40,34 +37,7 @@ resource "aws_route_table" "rt_dmz" {
   }
 }
 
-## Provisions NAT security group
-resource "aws_security_group" "nat_sg" {
-  name = "${var.stack_item_label}-nat"
-  description = "NAT security group"
-  vpc_id = "${aws_vpc.vpc.id}"
-
-  tags {
-    Name = "${var.stack_item_label}-nat"
-    application = "${var.stack_item_fullname}"
-    managed_by = "terraform"
-  }
-
-  ingress {
-    cidr_blocks = ["${var.lan_cidr}"]
-    from_port = 0
-    to_port = 0
-    protocol = "-1"
-  }
-
-  egress {
-    from_port = 0
-    to_port = 0
-    protocol = "-1"
-    cidr_blocks = ["0.0.0.0/0"]
-  }
-}
-
-## Provision VPC flow log
+## Provisions VPC flow log
 resource "aws_cloudwatch_log_group" "flow_log_group" {
   name = "${var.stack_item_label}-vpc-flow-logs"
 }
@@ -107,7 +77,7 @@ resource "aws_iam_role_policy" "flow_log_role_policies" {
         "logs:DescribeLogStreams"
       ],
       "Effect": "Allow",
-      "Resource": "*"
+      "Resource": "${aws_cloudwatch_log_group.flow_log_group.arn}"
     }
   ]
 }
@@ -118,5 +88,5 @@ resource "aws_flow_log" "flow_log" {
   log_group_name = "${var.stack_item_label}-vpc-flow-logs"
   iam_role_arn = "${aws_iam_role.flow_log_role.arn}"
   vpc_id = "${aws_vpc.vpc.id}"
-  traffic_type = "ALL"
+  traffic_type = "${var.flow_log_traffic_type}"
 }