+++ base/outputs.tf
@@ -15,11 +15,6 @@ output "rt_dmz_id" {
   value = "${aws_route_table.rt_dmz.id}"
 }
 
-## Returns ID of NAT security group
-output "nat_sg_id" {
-  value = "${aws_security_group.nat_sg.id}"
-}
-
 output "flow_log_id" {
   value = "${aws_flow_log.flow_log.id}"
 }